moving debug interfaces to debug package
